feature(aux-ts): add dataFeed address and public key (#22)
data feed user can read tables for order book and emit them into an
event queue. With data feed public key and address, the client can
simulate the reading operation on chain and get the resulted
events (market data) without paying any gas.

There are offical data feed for aux on devnet and testnet, and we use
the localnet deployment account as the data feed user.

If user of the library sets up an account for specific network and
saved it in the profile config, it will be used in lieu of the
official one.