feature(aux): reward distributor (#292)
feature(aux): reward distributor.

Reward Distributor is similar to `stake.move`, but is geared toward
primitive and lacks many of the functionalities of `stake.move`.
The intention is to distribute reward to users based on their share
amount, and when they got the shares. It is not intended to be used as
a standalone contract, rather, it is to be integrated into larger
modules where the reward is generated and distributed.

Also included is a `quote_coin.move`. Often, users want to perform
some calculations that involve `aptos_framework::coin::Coin`, such as
in RewardDistributor, estimate how much reward they can
get. `quote_coin` is a module that fully replicate the operations on
`Coin` (such as `merge`, `extract` etc), but anyone can arbitrarily
create and burn. This module can be used in place of coins so that the
same calculations can be performed. See `reward_quoter` for an example.